+++ Source/Scene/ModelExperimental/ModelExperimental3DTileContent.js
@@ -266,6 +266,10 @@ ModelExperimental3DTileContent.fromB3dm = function (
     content: content,
     colorBlendMode: tileset.colorBlendMode,
     colorBlendAmount: tileset.colorBlendAmount,
+    lightColor: tileset.lightColor,
+    backFaceCulling: tileset.backFaceCulling,
+    shadows: tileset.shadows,
+    showCreditsOnScreen: tileset.showCreditsOnScreen,
   };
   content._model = ModelExperimental.fromB3dm(modelOptions);
   return content;
@@ -295,6 +299,10 @@ ModelExperimental3DTileContent.fromI3dm = function (
     content: content,
     colorBlendMode: tileset.colorBlendMode,
     colorBlendAmount: tileset.colorBlendAmount,
+    lightColor: tileset.lightColor,
+    backFaceCulling: tileset.backFaceCulling,
+    shadows: tileset.shadows,
+    showCreditsOnScreen: tileset.showCreditsOnScreen,
   };
   content._model = ModelExperimental.fromI3dm(modelOptions);
   return content;
@@ -324,6 +332,10 @@ ModelExperimental3DTileContent.fromPnts = function (
     content: content,
     colorBlendMode: tileset.colorBlendMode,
     colorBlendAmount: tileset.colorBlendAmount,
+    lightColor: tileset.lightColor,
+    backFaceCulling: tileset.backFaceCulling,
+    shadows: tileset.shadows,
+    showCreditsOnScreen: tileset.showCreditsOnScreen,
   };
   content._model = ModelExperimental.fromPnts(modelOptions);
   return content;